Add the Hero Section in Home Page
