+++ job-siege/src/HomePage/HeroSec.css
@@ -1,0 +1,72 @@
+
+.hero-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    padding: 50px;
+    max-width: 1200px;
+    margin: 0 auto;
+    margin-top: 60px;
+  }
+  
+  .hero-image {
+    flex: 1;
+  }
+  
+  .hero-image img {
+    width: 100%;
+    max-width: 500px;
+    height: auto;
+    border-radius: 10px;
+  }
+  
+  .hero-text {
+    flex: 1;
+    padding-left: 30px;
+    font-weight: bold;
+  }
+  
+  
+  .hero-text p {
+    font-size: 2.5rem;
+    line-height: 1.6;
+  }
+
+  @media (max-width: 880px){
+    .hero-text p{
+        font-size: 1.5rem;
+    }
+  }
+  
+  /* Media queries for responsiveness */
+  @media (max-width: 768px) {
+    .hero-container {
+      flex-direction: column;
+      padding: 20px;
+      margin-top: 80px;
+    }
+  
+    .hero-text {
+      padding-left: 0;
+      padding-top: 20px;
+      text-align: center;
+    }
+  
+  
+    .hero-text p {
+      font-size: 1.5rem;
+    }
+  }
+  
+  @media (max-width: 480px) {
+  
+    .hero-text p {
+      font-size: 1rem;
+    }
+
+    .hero-container {
+        margin-top: 90px;
+    }
+  }
+
