+++ job-siege/src/HomePage/Home.css
@@ -1,161 +1,7 @@
-
-body {
-  margin: 0;
-  font-family: "Poppins", sans-serif;
+.entire{
+    font-family: "Poppins", sans-serif;
 }
 
-.lines {
-  height: 62px;
-  width: 1px;
-  margin: 0px 25px;
-}
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: rgb(255, 255, 255);
-  padding: 10px 20px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.nav-left,
-.nav-middle,
-.nav-right {
-  display: flex;
-  align-items: center;
-}
-
-.nav-right {
-  justify-content: space-between;
-}
-
-.profile-pic {
-  height: 30px;
-  width: 28px;
-  border-radius: 50%;
-  margin-right: 15px;
-}
-
-.username,
-.app {
-  color: black;
-  font-weight: bold;
-}
-
-.app {
-  margin: 0;
-  padding: 0;
-  font-size: 40px;
-}
-
-.app-logo {
-  height: 50px;
-  width: 50px;
-  margin-right: 20px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.5);
-}
-
-.nav-link {
-  color: #535151;
-  text-decoration: none;
-  font-weight: bold;
-  transition: color 0.3s;
-}
-
-.nav-link:hover {
-  text-decoration: underline;
-  color: #975842;
-}
-
-/* .... */
-
-.mobile {
-  display: none;
-}
-.hamburger {
-  display: none;
-  cursor: pointer;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: 24px;
-}
-
-.hamburger .bar {
-  width: 30px;
-  height: 4px;
-  background-color: rgb(0, 0, 0);
-  margin: 3px 0;
-}
-
-.menu {
-  display: none;
-  flex-direction: column;
-  align-items: center;
-}
-
-.menu.open {
-  display: flex;
-}
-
-.menu-item {
-  padding: 10px 0;
-  color: white;
-  text-decoration: none;
-  font-weight: bold;
-  transition: color 0.3s;
-}
-
-.menu-item:hover {
-  text-decoration: underline;
-  color: #975842;
-}
-
-@media (max-width: 880px) {
-  .mobile{
-    display: flex;
-    align-items: center;
-  }
-
-  .app{
-    font-size: 25px;
-  }
-
-  .username{
-    display: none;
-  }
-
-  .menu{
-    display: none;
-    flex-direction: column;
-    position: absolute;
-    top: 60px; 
-    right: 20px; 
-    background-color: #535151;
-    border-radius: 5px;
-    z-index: 1;
-    padding: 10px; 
-  }
-
-  .menu.open{
-    display: flex;
-  }
-
-  .desktop{
-    display: none;
-  }
-
-  .lines{
-    margin: 0px 18px;
-  }
-
-  .app-logo {
-    height: 40px;
-    width: 40px;
-  }
-
-  .profile-pic{
-    margin: 0;
-  }
+.header{
+    margin-top: 50px;
 }
